dhcp: allow overriding the BOOTP filename via subclassing
This enables subclasses to implement their own policy for selecting
which filename should be returned to the PXE client.

This refactoring also fixes the inconsistency where a TFTP or HTTP
link would be returned in get_pxe_filename() when not using the iPXE
mode and the architecture option (93) would be missing.