dhcp: find the options in self.options, not the client's lease
Up until 317d02d9449d, the DHCP options for a client were found in the
client lease. This however was changed in favor of using a dictionary
instead.

I agree with the change, since I don't see why we should store options
in the lease, but it failed to update the code that was selecting the
best filename per architecture.

This code was further-disabled by 0d1b0d5d2a2f ("pypxe/dhcp.py: avoid
exception when leases have no options"), by not crashing when the
code would have been used.

Let's fix the original regression properly, until we propose a new
way of storing the preferred file names.

Fixes: 317d02d9449d ("Added exporting leases to json. Fixes #71")